requirements.txt
streamlit
openai-whisper
argostranslate
argostranslate-tools
